move organization views to their own file
